Python GUI
Added a python gui for message generation works on version 2 or 3.
Default directories in message_definition.xml file

+++ message_generator/generate/generate_proto_wrapper.py
@@ -5,19 +5,27 @@ import shutil
 from .proto_wrapper_src import *
 from .proto_wrapper_include import *
 
-def generate_proto_wrapper(directory, include_extension, src_extension):
+def generate_proto_wrapper(directory, include_extension, src_extension, deleteNew):
     """
     Function will create source & header folders for c# wrapper and files for c# wrapper.
     If files or folders already exists function will delete them.
     All C# files are auto generated in desired directory. No file are copied.
-    """       
-    if os.path.isdir(directory):
-        shutil.rmtree(directory)
-        print("Deleted directory: " + directory)
+    """
 
-    if os.path.isdir(directory) == False:
-        os.mkdir(directory)
-        print("Created directory: " + directory)
+    if(deleteNew == True):
+        if os.path.isdir(directory):
+            shutil.rmtree(directory)
+            print("Deleted directory: " + directory)
 
-    generate_proto_wrapper_include(directory, include_extension, src_extension)
-    generate_proto_wrapper_src(directory, include_extension, src_extension)
+        if os.path.isdir(directory) == False:
+            os.mkdir(directory)
+            print("Created directory: " + directory)
+            
+        
+        generate_proto_wrapper_include(directory, include_extension, src_extension)
+        generate_proto_wrapper_src(directory, include_extension, src_extension)
+    
+    else:
+        
+        generate_proto_wrapper_include(directory + "/include/", include_extension, src_extension)
+        generate_proto_wrapper_src(directory + "/src/", include_extension, src_extension)
